pipeline/requirements.txt
supabase>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
shapely>=2.0.0
geopandas>=0.14.0
numpy>=1.26.0
